NITs cleaning
